Restructure packages for online banking CZ
COAND-497
