+++ online-banking-cz/src/main/java/com/adyen/checkout/onlinebankingcz/OnlineBankingCZComponent.kt
@@ -18,6 +18,7 @@ import com.adyen.checkout.core.log.LogUtil
 import com.adyen.checkout.core.log.Logger
 import com.adyen.checkout.onlinebankingcore.OnlineBankingComponent
 import com.adyen.checkout.onlinebankingcore.OnlineBankingDelegate
+import com.adyen.checkout.onlinebankingcz.internal.provider.OnlineBankingCZComponentProvider
 
 class OnlineBankingCZComponent internal constructor(
     delegate: OnlineBankingDelegate<OnlineBankingCZPaymentMethod>,
